fdo#42924 don't ignore empty TableCell Element, also check if it's a header
Checking whether an empty TableCell Element is a header row is now in 2 cases:

* | e | h | h | h | h | ... => The empty cell should be a header
  This case occurs when this row is the first row in the table.

* | h | p | p | e | p | ... => The empty cell should not be a header
  This case occurs when this row is not the first row in the table,
  but the first column is the row header.

where | is a column seperator, e is an empty cell, h is a header cell
(which has role="tablehead" attribute), and p is not a header cell.

Note that parsing occurs left-to-right, so isTableHeader depends on
the last TableCell Element in that row. I assume that if the last element
is a header, that row should be a header row.

Currently this code gives correct behaviour, but checking whether a row
is the first row might be more correct.

Signed-off-by: Petr Mladek <pmladek@suse.cz>