tdf#89437 Add note, not all languages have autocorrection of ordinal suffixes
Not all languages have the convention of printing ordinal suffixes as superscript.
For English auto correction of 1st to 1^st can be correct in Swedish auto
correction of 1:a to 1^:a is just plain wrong and LibreOffice does not do it anymore
so lets add a short note about it in the help.

Change-Id: Ic4c9273b74576d2a8d1e0c8ac3adbe4af40185fc