tdf#94553 fix nested lists in wikihelp
The problem was that only the first element in a nested list
was getting the extra '#' character.  Added a re.sub statment to ensure
all elements inside a nested list have this to ensure proper display
Also backing out temporary/FIXME changes in the xhp files.
The run is now free of warnings with the desired changes showing up
for 'Using Gluepoints' page, which has a nested list

Change-Id: I4009408ee11e191cbd5e1560ca1d08f3e7b5c630