help-to-wiki shell call replaced with a function
shell call to run getalltitles has been replaced
with a function call.

Also added a new option to save the title file
alltitles.csv  If this file is needed, to continue
to generate this, please add -t to the run.
Or else, please remove it so that a stale file
is not kept around

Change-Id: I2902243df59d415fb313efa7d4132b0190658fa3