Fix ahelps for Mathematical functions
In the code there is usually a shorter identifier macro
defined which is replaced via a header file to a real helpid.
In the help for some reason this shorter identifier is used,
which might have worked some time ago but it does not today.
I'll mass replace the 1000+ such instances until 5.4.

Change-Id: I81d4d7d65f695e30f37081b75899a3f1cdf8cdd3