tdf#60874 (partial) Show correct page from "Protect Document" window
The bookmark for the Password entry was used on the wrong page.
This needed to be moved over and the entry description set up
as extended help.

A new help ID was needed to be added to the Confirm entry
with an accompanying extended help. Hidden, as this page
is rather a howto and not a window description (thankfully).

Change-Id: I6583f8708cd4c0b0891ead7067ce486153dd2be1