helpcontent2: fix incremental rebuild of *.ilst files
They only depend on the perl script so won't get rebuild if images are
added or removed in source/media/

Use the same hack as in postprocess, make the targets PHONY so they
always get rebuilt.

Also fix a stray \ that caused the wrong output "[PRL] " to be printed
too early.

Change-Id: I03c8696591cf47395211feaed360fbd5642733a0