TDF#97745 Add support custom search in helponline
Added some language support for the helponline global search

Each custom search engine of google has a unique token, which
configuration is handled by CSE owner.

The CSE attempt to seach in helponline in the page language. If a page
is already indexed, it shows what it has. If not, pages indexed from the
*.libreoffice.org domain and wiki.tdf.org are shown.

The CSE prevents display of indexed pages of other suites.

Change-Id: I133e1a7af867c2380ccb1f20c483af63337a58b9
Reviewed-on: https://gerrit.libreoffice.org/42218
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>
Tested-by: Olivier Hallot <olivier.hallot@libreoffice.org>