Use list.js for Index listing and search
List.js - http://listjs.com/ - allows us to paginate the results
so the height does not get out of hand.
Fuzzy search would be cool: http://listjs.com/docs/fuzzysearch/
...but it gives too big of a perf hit with our 5700 items.
Might try to bribe the creator to look into perf.
Thanks a lot to David Tardon for the makefile solution and to
Christian Lohmaier for playing a rubber duck.

Change-Id: I359eed541470ccaa309b0b6ff5d809a796d9befd
Reviewed-on: https://gerrit.libreoffice.org/48335
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>
Tested-by: Olivier Hallot <olivier.hallot@libreoffice.org>