tdf#115524 Some branch bookmarks are not working 1
With .ui dialogs, the Help system sends specific URL
associated to the higlighted widget in the .ui dislog.

Not every help page associated to the dialog has
<bookmark branch="hid/<path to the dialog/widget"/>
for all widgets of the dialog. As consequence, some URLs
were not addressed and the default module entry page
was displayed, instead of the dialog help page.

The solution is to define a dumy bookmark '@@nowidget@@' to the
help page associated to the dialog, and if the widget bookmark
is missing in the help page then use the dummy target,
thus offering a help page with the right contents anyway.

This patch covers help pages for smath.

Change-Id: I40b592271b0c336d4d3bacc33d2f3d19107ae702