Default Add value for Optimal Column Width is 2 mm, not 0.1 in
See ScCellShell::Execute implementation [1] which passes STD_EXTRA_WIDTH
(defined as 113 twips = 2 mm) to the dialog as the default value.

The text here (that was 0.1 in) is translated to other languages (incl.
those using metric units) verbatim (e.g. Russian translation); or if a
translation team would want to convert to metric, there is no indication
that the 0.1 in is just an approximation of an exact metric value.

[1] sc/source/ui/view/cellsh3.cxx:693

Change-Id: I1710588ee40840d85eeb89aaeee135a8aa0a942f