Improve donation link in Help online
redmine#2815

Addressed responsiveness: for screen width < 960px the
button goes to the bottom of the screen and the header is
not fixed anymore.

Artists are invited to improve button rendering.

Button max width is 300px, the en-US text and the translated
text should fit in 2 lines of 300px each. Text is centered.

Change-Id: I51860be2ad998e85f247118e44b21af3e2efa060