tdf#102265 update Impress Help as Format - Page was moved
This is a complex task and the intertwined nature of Draw and
Impress does not help.

In the patches for tdf#84258 .uno:SlideSetup was created as an
alias for .uno:PageSetup. The corresponding menu items are
Slide - Properties in Impress and Page - Properties in Draw.

Currently some Draw help pages refer to the simpress/01/01180000.xhp
file. This file needs to be cloned under sdraw and the contents
of both files along with all references to them have to be
brought up to date.

The Impress Slide menu documentation needs to be fleshed out in
simpress/main0117.xhp and new files have to be created for the
individual menu item references.

The Impress Format menu documentation simpress/main0105.xhp
needs a big overhaul. The same is true for sdraw/main0105.xhp.

The Help for Draw's Page menu has to be created.

To find the interrelations, I use
find /path -name '*.xhp' -exec grep --color=always -Hi 'string' {} +
or with -type f instead of -name to get maximum coverage.

Change-Id: I5d130846da6640aeebab758c0f6dae2c78f9e355