Donation button and sticky header for Help
Task: https://redmine.documentfoundation.org/issues/2815

Header was made sticky on non-mobile by using position: sticky.
This was the only way to make it work with the grid without
carving magic numbers into stone. IE11 ignores it, but so what?

Donation text was made shorter.

The Google search and the donation button were wrapped in a
container to make it easy to haul them all over the place.

Lots of tweaks were made to the grid areas and the styles of
the bottom elements to make everything fit nicely in the new
markup order. The result was tested on IE11 to make sure it
degrades gracefully.

The Displayarea element had a fixed width (my old mistake),
which made the min-width: 960px layout look very uncool.

Change-Id: I131a1e82290bc3381c938acb4dc9dbc408fb2ba7
Reviewed-on: https://gerrit.libreoffice.org/68890
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>
Tested-by: Olivier Hallot <olivier.hallot@libreoffice.org>