Show module & lang navigation on mobile
Module & languages menus were not shown on screen widths < 960.

Changed module & languages menu background to span the whole
width, when on mobile.

An arrow function was used in help2.js, which means it was broken
on IE.

Moved all src .js files to head and used defer for all.

Moved inline Piwik & system info JS into help2.js and added a
test for onlineness targeting the first meta element in the body.

Change-Id: Ic1a928b53ce9029f9f655117977cd511b7e3246d