Create XSLT replacement table for missing icons
Missing UI icons have a replacement table in
/core/icon-themes/colibre/links.txt

Help must also use this replacement table when Help
pages have missing icons. The previous approach was a manual
creation of a XSLT for the replacement.

The patch moves the replacement table creation inside the
Custom makefile, making it automatic and carrying changes
of the original links.txt source.

Change-Id: I7f834fa03bbd85f329efbec43e5c51cc86d51586
Reviewed-on: https://gerrit.libreoffice.org/70080
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>