tdf#124198 Impress & Draw ImageMap menu entry was moved to Tools
Also renamed the file describing Edit Menu to edit_menu.xhp.
Most of the files are touched to update the links due to rename.

There is no context menu entry for ImageMap, so removed mention
of it.

Note that in Writer and Calc, ImageMap remains under Edit, because
it is only enabled when an image is selected.

Change-Id: Ic57ee11dd2ab9c637f00e5328f1575c323fb8c43
Reviewed-on: https://gerrit.libreoffice.org/71397
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>