redmine#2555 replace Help online Google search
This is the helpcontent2/ part of the patch.

Add xapian-omega search to online Help.

The patch replaces Google custom search with xapian-omega search.

A new build key is introduced.

--with-omindex=server : Localizes and adds the xapian result page
template, adds the xapian form to each Help page.

--with-omindex=noxap : do not localize the result template and
do not add a form in the Help page.

--with-omindex= will force Online Help build.

Default is noxap

NOTES:
- xapian-omega databases are built in the server.
- searches returns results only on localized Help pages, avoiding same
resulis in many languages.

TODO:
- Tweak the xapian-omega result page CSS and markup.

Change-Id: I3dc2b79c4111f61fd5fea01b1f8c95bfacab1e8f
Reviewed-on: https://gerrit.libreoffice.org/79366
Tested-by: Jenkins
Reviewed-by: Guilhem Moulin <guilhem@libreoffice.org>
Tested-by: Guilhem Moulin <guilhem@libreoffice.org>