redmine#2555 replace Help online Google search
This is the helpcontent2/ part of the patch.

Add xapian-omega search to online Help.

The patch replaces Google custom search with xapian-omega search.

A new build key is introduced.

--with-omindex=local : Builds the xapian omega index for each language,
localize the result page and add xapian form in Help pages.
This option should be used if the target server cannot build
the xapian databases. The data volume of the Help online
will increase five fold approximately. Use with care.

--with-omindex=server : Localizes and adds the xapian result page,
adds the xapian form to each Help page but does not build the xapian
databases, which is left to be done in the server directly.

--with-omindex=noxap : do not build the databases, do not localize
the result template and do not add a form in the Help page.

--with-omindex will force Online Help build.

Default is noxap

NOTES:
- xapian-omega is built by $productversion and by $language
- searches returns resuls only on localized Help pages, avoiding same
resulis in many languages.
-xapian-omega databases location are hard-coded in server. For
Debian familiy it is in /etc/omega.conf and as example:
a) set database_dir /var/www/html/$(productversion)
b) set template_dir /var/www/html/$(productversion)

TODO:
- Tweak the xapian-omega result page CSS and markup.
- create a xapian-omega helper script to be run in server
- Set the omindex parameters for target server.
- add xapian-omega as build-dep, if local= is used

Change-Id: Ic97739d1a0a705d88ceab4e1d2fe17230e938a27