Related: tdf#82375: Rewrite descriptions for the last argument of HLOOKUP/VLOOKUP
That parameter is not about telling the *functions* if the search column in
the array is sorted or not, but about different semantical treatment of the
array, i.e. whether it's the array of values or sorted ranges. The wording
"Sorted" and "indicates whether the first column in the array is sorted in
ascending order" is misleading, not providing the idea behind the argument,
and create an impression that all sorted array lookups must have it ser to
TRUE.

Change-Id: I795c29f3c7d5632db6f72f7906bfdacaa372832e