tdf#130064 Help page for legacy list style numbering in Writer
ODF has two different methods for positioning lists. LibreOffice Writer
has always used one method ('label-alignment'), and the relevant
controls are described in /text/shared/01/06050600.xhp for the
Numbering Style Position tab dialog.

If LibreOffice opens a document produced with the other method
('label-width-and-position'), then different controls will be shown in
the Position tab of the Numbering Style dialog. This help page provides
information about these controls. The controls are called "legacy" to
have a simple, meaningful label (where there is no standard or
familiar label). These controls are only "legacy" in relation to
LibreOffice Writer, but they remain legitimate controls, and this help
page describes how the controls appear in the Position tab
when documents are opened that use this other method.

-based on embedding or moving content from /text/shared/01/06050600.xhp

 (WIP)

Change-Id: I2942f2daa9ffbab92dffd8f333e7d855e17f6fda