Update patch set 3
Patch Set 3: -Code-Review

> I think you can edit the patch yourself and complete the missing part, in case you work directly in gerrit, or copy/paste the contents in your favorite text editor (actually the stub was created in the xhpeditor).

I thought so too -- but when I opened the position tab page that was just merged, it shows an older version.  And there was no "rebase" option.  So I figured a way around this situation was for you to merged this patch, and then I could "load" this new file.  (I should be able to finish it today -- so the stub would not sit empty for long.)

Patch-set: 3
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=0