Update patch set 5
Patch Set 5:

There was a patch recently for the Case labels in the Font Effects in Writer.  I started to update the labels on the Font Effects page -- but noticed a need to update more, hence this patch.  Have not changed any text -- beyond corrections to align the commands and controls on this page with the actual dialog box.  Will assume it is ok, unless I hear otherwise.

Patch-set: 5
Label: Verified=0