Update patch set 8
Patch Set 8: Code-Review-1

Ok - I think that I have sorted this out now.  I can see that it is Writer Style is the only dialog that does not have the "Transparency" control -- so I have "repaired" the <switch> so that it shows the Transparency control for Calc, Draw, and Impress, but not for Writer. 

(this is still wrong in relation to "Format Character" in Writer, which does have Transparency - but I guess if the Help from Writer Style pages go to a different page (that embeds everything from here, except for Transparency), then it would match.  

So the only question that seems still particularly relevant is the part in "Font Color" that seems to be for "highlighting" (and probably should be removed).

Sorry for the chaos - was hard to get an understanding of the variations.

Patch-set: 8
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1