Update patch set 3
Patch Set 3:

> The design was to have the <base> relative to the Help web root.

I see, but why is that not changeable?

> If online, then URL is  http://<server_root>/<version>/<pages>.html

The URI should not change as the result of this change.

> ...
> We use the version number for online because we want to have all online Helps available, and in synch with the software, including media. Users of old versions (post 6.0) still have the corresponding online help of their LO version.

Of course! And the change does not change that. The change only makes the links in the page itself relative to one level higher. When you are inside server/6.3, any link inside will still lead to server/6.3/otherpage, just does not need to explicitly spell that number... what do I miss?

Patch-set: 3