Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> please don't add empty paragraphs marked for translations like

Christian -- thanks for pointing this out.  They were never intended to be in the patch. It was an oversight.  A template was pushed for me in this patch -- and I did not notice the things that you pointed out were in the template.  Of course I will clean this up.

Patch-set: 4