Update patch set 11
Patch Set 11: Code-Review-1

(1 comment)

@Olivier, I could see that the "Autoformat for Tables" page has a similar structure to this page -- first explaining how to apply table styles and then explaining the elements.  This structure could be a simple ("temporary") solution for this page. 

If so, I could make a proposal (i.e. make a patch with the changes) -- but it would be better to push this current patch (with the change of "section id" and the other "simple" changes, where the "Table Style" addition is also added.  There are no "loose ends" now.).

And then I could make a new patch -- making the bigger changes.  Or -- alternatively -- the things that I would change are already noted in the unresolved comments, and could be included in this patch.  Just looking for a way to get this finished.  Thanks.

Patch-set: 11
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1