Update patch set 5
Patch Set 5: Code-Review-1

You need to pick the bookmarks/HID I specified in my abandoned patch and place them correctly in your page. As it is, it will not fix tdf#130733. Each bookmark is associated to a widget in the dialog, and is obtained from the .ui file of the dialog. The bookmark is the link target when user select a widget and press F1.

See https://gerrit.libreoffice.org/c/help/+/88867

Please do not tell user to press OK or Cancel, it is too obvious.

Patch-set: 5
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1