Update patch set 11
Patch Set 11: Code-Review+1

(1 comment)

This patch should be ready to go.  Solved/clarified the transparency problem.
There are no transparency controls for Paragraph Style and Character Style (tdf#130456), so I added a <note> inside of a <switch> for Writer.

The only other issue that could be addressed is that I believe lines 67-75 should be deleted, because they do not belong here. They are about "highlighting" and not about "font color".  (but it was something that was already there -- so not necessary for the main purpose of this patch, to update the help for "Case".)

Patch-set: 11
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=+1