improve links to module-specific help pages for Styles window
Change the present method of linking the "Styles" header for each
module to a generic page with an activated link to the relevant
styles help page to go directly to that page.

Each of specific pages calls "text/shared/02/01230000.xhp#stylist"
This patch moves the switch in this page to be applied to the Styles
header.

Change-Id: Icb846f44d5eb61e8a42612c0d8e05de2c959f7d2
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/89786
Tested-by: Jenkins
Reviewed-by: Seth Chaiklin <sdc.blanco@youmail.dk>