Update patch set 1
Patch Set 1:

> We don't have a "Styles window" in GUI, we have only "Styles" section (or deck) in Sidebar in all modules

Hi Roman.  Thanks for your message. Some help pages refer both to "Styles Window" "Styles deck" on the same page. When I search the source for the help pages, then I see "Styles window" many times (and "Styles deck") only a few times (and in general "deck" is not used in the help pages).  

I appreciate that this may not be the right "technical" language -- but if most of the help pages are using Styles window, then, as noted in the commit message, it seems better to be consistent (to reduce confusion). Happy to hear your opinion about what to do in this situation.

Patch-set: 1
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>