Update patch set 3
Patch Set 3:

> 
> What do you think?

I think such changes are cosmetic and affect not only the translators, but also the guides published by the community. I just reviewed the Getting started guide and found deck in use. Actually the deck is contained in a panel of the sidebar, so, window is not the best word. Deck is to sidebar what tabs are for dialogs. 

Before patching a change in nomenclature, please open a bug on the subject and let the developers and UX specialists (htieze) advise. Besides, "deck" may be also used in the UI (not checked)

Patch-set: 3
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>