Update patch set 3
Patch Set 3: Code-Review-1

Styles used to be in a window back in 2012, when we didn't have the Sidebar yet. Try launching 3.5.0 and hitting F11. Sure, we can undock the Sidebar into a separate window, but it's still a Sidebar with various decks in it.

Rather change the occurrences of "Styles window" to "Styles deck". This is just another case of hopelessly outdated Help content.

Patch-set: 3
Reviewer: Gerrit User 1001613 <1001613@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1