Update patch set 4
Patch Set 4:

Hi Oliver,

Thank you for the reviews!

I just wanted to say that you don't need to waste your time with fixing commits in the "Merge Conflict" state - I can take care of fixing them, after which they can be reviewed and merged.

As to the addition of the "tdf#132643" to the commit message - I intentionally only added it to one of the commits, so that it appears in the Bugzilla bug, to let people that there is some progress. I didn't add it to the rest, to not spam the bug report - otherwise the useful comments there will be drowned in the commit messages, and may be overlooked. Instead, I set the same topic "tdf132643" for all commits, so they can be easily found. But if you prefer, I can add the "tdf#132643" to all future commit messages as well.

Patch-set: 4
Reviewer: Gerrit User 1001316 <1001316@6622c302-437a-49ad-a91d-55ba7dcf1615>