Create patch set 2
Patch Set 2: Published edit on patch set 1.

Patch-set: 2
Subject: The syntax of the GROWTH() function is still not quite correct in the Help. GROWTH() is special in that the 2nd and 3rd parameters are not only optional but may also be specified empty/omitted, in which case they are set to defined values. This makes the syntax a bot more complex but it can be checked against the specification in ODF V1.2.
Commit: 69cb695e553271452a5ff95558bbaee1c75f909f
Tag: autogenerated:gerrit:newWipPatchSet
Groups: f0047016243df9f12df294b3a4e8888366af7ebc