Update patch set 1
Patch Set 1:

The ahelp tag is not implemented in the current Help. However, a better solution (to be discussed) is to move the ahelp contents to the dialog itself (there is room for tooltips in the widgets). The ahelp will be kept in the Help page because it is part of the Help text.

Advantage: fix the tooltip in dialogs. 
Disadvantage: unless done by a script, there will be a retranslation effort.

To see the tooltip in Glade, open a ui file, select a widget and look at the "common" property tab, there is a tooltip field.

Patch-set: 1
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>