Adjust OFFSET() note to reality
It was never necessary to enter OFFSET() as an array formula if a
cell range reference is returned. Whether to use an array formula
entirely depends on how the resulting range(s) are to be used as
arguments to calling functions.

Change-Id: Icb771c680f1b9c076017ebd2c78f440529ecc1b2
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/97620
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins
(cherry picked from commit c8683685e228f1daae3d039886245b6dfea6d138)
Reviewed-on: https://gerrit.libreoffice.org/c/help/+/97580
Tested-by: Eike Rathke <erack@redhat.com>