Update patch set 1
Patch Set 1:

Hi Adolfo, László,
I talked with Sophie about this patch and we agreed to revert it since it was submitted after string freeze and it's not a critical issue.
Unless there is a strong reason not to revert it, I would like to do it before RC2 so please speak up if you have a reason not to revert it.
Thanks

Patch-set: 1
CC: Gerrit User 1000007 <1000007@6622c302-437a-49ad-a91d-55ba7dcf1615>