Update patch set 1
Patch Set 1:

I do not mind you leave it, just let's (once and for all) set the procedure straight.
If there is some urgent UI/help string patch/addition, it *must be properly communicated* with the l10n head so l10n teams are informed on time to give it proper attention. And of course, if it is just a string or two at the time of RC1, that is not as much as a problem as 100 strings after RC2 ...
So, I already localized these few strings in my system and my translation is already checked in, so I do not see need to revert the patch. Also this is a minor patch.

Sophie can probably say her opinion on this as the l10n lead.

Patch-set: 1