Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Yeah, let’s talk about miscommunication. ESC supossedly agreed to extend the beta period a couple of meetings ago, and now we’re on RC?
> 
> Can’t we let this stay?

Could you please share the link to the minutes mentioning that? I attend the ESC minutes regularly and I don't remember it was ever said. OTOH, we did agree to have a second beta in order to fix the issue with mac langpack installers. Maybe you misunderstood it ?

Anyway, if translator are happy about it, I'm fine with it not getting reverted but please, be more careful next time

Patch-set: 1