Made Clear Direct Formatting help match reality
The Clear Direct Formatting help text incorrectly states that
this function will remove character styles on Writer, which
is not true. This is causing confusion as reflected in the
following bugs:

Bug 132702 - CLEAR DIRECT FORMATTING: Wrong information in LO Help
- https://bugs.documentfoundation.org/show_bug.cgi?id=132702

Bug 132469 - Clear Direct Formatting char does not work
- https://bugs.documentfoundation.org/show_bug.cgi?id=132469

as well as in the following question from Ask:

Why does direct clear formatting not clear character style
- https://ask.libreoffice.org/en/question/51776/

This was an old behavior pre-2011 before I converted "Default
Formatting" into "Clear Default Formatting". The help text was
also updated by me back then, but I am observing that it was
done incompletely, as it seems like I changed the topic title
but did not update the text about the actual behavior:

https://git.libreoffice.org/help/+/7d23506818f48f534622f83343a79083105672da%5E%21/

Additionally, while analyzing this issue, I found that people
get confused as to how to remove remaining formatting that is
still applied as a character style. Sometimes they don't know
that a character style is still applied. Thus, I am adding a
note about it.

The Design Team or other developers may later decide that it's
better to proceed in a different way, but, for now, it's
important to state the truth about how this function works.

Change-Id: I2a522c68630d64c21eabe9af61ace8a66d633ae2