Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> And I, once again, would like you to compress your SVGs with svgcleaner, as SVG artists do in the core repository. (And any text in SVG images should be traced with Inkscape; otherwise its rendering will differ across OSes)

Adolfo: you need to give some tips. I tried with svgcleaner and SVGOMG. Both produced a messed up result (note: you have to view in the browser, Inkscape is messing it up by default). With svgcleaner, I had to temporarily cut out the <style> element, because it refused to clean the file with it present.

https://github.com/RazrFalcon/svgcleaner/
https://jakearchibald.github.io/svgomg/

Patch-set: 5
CC: Gerrit User 1001613 <1001613@6622c302-437a-49ad-a91d-55ba7dcf1615>