Update patch set 4
Patch Set 4: Code-Review-1

(6 comments)

Here is attempt to patch help for "Document" fields.  (Most of the page was written at least 8 years ago, so there are all kinds of problems.) For now have attempted a "minimal solution" -- to delete the things that should not be here and add one missing item. (these are all in the comments). 

Would also be interested to try to improve the presentation at the end, if I may.  (Made one suggestion in a comment for Date and Time).
And have filed a bug report because some of the items in "Format" should be in "Select" (which should make it easier to write help).

Patch-set: 4
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1
Work-in-progress: false