Update patch set 7
Patch Set 7:

(3 comments)

When I look in the history of this patch, then it seems that Patch set 1, and not the last (patch set 7) was committed.

Olivier committed this patch on 5 Oct, and I can see the revision tag:  # fc8fdf63 
in the source, but it also says 1 Oct and not 5 Oct.  Here are the links to show the problem.

https://opengrok.libreoffice.org/history/help/source/text/shared/00/00040500.xhp
https://opengrok.libreoffice.org/history/help/source/text/shared/01/05040200.xhp 

I would like to have the latest patch set here applied -- but I do not know what/how to do it
in a good way.  Please can one of you help!

Thanks

Patch-set: 7