Update patch set 4
Patch Set 4:

(2 comments)

To Olivier and others -- I am having a problem with <switch> and "CALC".  The two comments identify two places (in the merged) code, where I believe that the content should be shown on the CALC pages -- but they are not -- and I am not able to identify why.  

Here is the link to the relevant page: 
https://help.libreoffice.org/7.1/en-US/text/shared/01/05230100.html?&DbPAR=CALC&System=WIN

Thanks for your help with this mystery (for me).

Patch-set: 4