Update patch set 7
Patch Set 7:

> Patch Set 7:
> When I look in the history of this patch, then it seems that Patch set 1, 
> and not the last (patch set 7) was committed.

Maybe the problem is/was with my understanding and not with the merged code.

Does opengrok record the data when the first patch set was made (and not the date when the final patch set was committed)?  That is what seems to be the case here.

I was confused that the date of the merge was 5. oct, but it appears as 1. oct in opengrok.
> https://opengrok.libreoffice.org/history/help/source/text/shared/00/00040500.xhp
> https://opengrok.libreoffice.org/history/help/source/text/shared/01/05040200.xhp 

But I can see -- in the online help -- that the "latest" patch from this merge is appearing as expected, and it looks right in the merged files, so I guess everything is ok.

If this is the explanation (that the date of the first patch set is reported, and not the date of the final patch set that is merged) -- then I am happy that the only problem was my ignorance and not a material problem with the merge.

Patch-set: 7