Update patch set 4
Patch Set 4:

Progress with the "switch" problem.

Here is simplest case. (verified in XHP Editor)

Problem seems to be that "switchinline" disrupts the switch.  

Please try the switch (at the end) in XHP Editor.  It is only a "default" -- so all modules should see its contents.  But DRAW sees nothing.  (changing the select to another module in the caseinline will change the problem to another module -- i.e., not specific to DRAW).

This "behavior" explains why I had problems with the "Calc" switch -- 

But now the critical question:  Is this expected behavior -- that a switchinline cannot be used inside a case of a switch (even though it validates as XML?)

-------- clip out here -----

<switch select="appl">
<default>
<h2 id="hd_id3153525" xml-lang="en-US">DEFAULT</h2>
<paragraph role="paragraph" id="par_id971603147148774">
DRAW will not see this, even though caseinline appears after this paragraph.</paragraph>

<paragraph  role="paragraph" id="par_id591602522244128"><switchinline select="appl"><caseinline select="DRAW">DRAW should see this -- but won't!</caseinline><defaultinline>DRAW Will Not See This</defaultinline></switchinline></paragraph>

</default>
</switch>

----------  end of test fragment -------------

Patch-set: 4