Update patch set 2
Patch Set 2: Code-Review-2

The "revised" Navigator is highly sophisticated, because the Categories have context-sensitive menus, which change for each category.

There are bugzilla requests for documentation -- but it is not clear (to me) how to approach this in the Navigator help page.

The current help page is updated now to correspond to the current layout of Navigator.

The "problem" is that most of the context menu items do not appear on the help page already, plus many of them do not really need to be explained.

The "patch proposal" here starts to make one way to approach the idea. Not complete in any way -- just trying to give an impression about an approach that could be developed further (or dropped).

This is a case of needsDocEval.  

Will not work further with this patch without some kind of agreement about how to approach this.

Patch-set: 2
Reviewer: Gerrit User 1000855 <1000855@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1000321 <1000321@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1001613 <1001613@6622c302-437a-49ad-a91d-55ba7dcf1615>
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-2
Work-in-progress: false