Update patch set 4
Patch Set 4:

Hi Ilmari - I hope you can help me with a small problem.  A recent German id update changed "objekt" to "object" in the Glossary and has created a "merge conflict" for me.  I tried "removing" that one file from my patch, but that did not solve the problem.

According to the tooltip for Merge Conflict, it says "download the patch, run "git rebase master" (resolve conflict) and upload.

In principle I have already resolved the conflict (by removing the offending file) -- but I am not able to run git locally.  I have downloaded the patch set (as a .zip), which I could send.  Or maybe you can get the right patch set here -- it is still an "edit" set -- here is the SHA number: fcfdd2b6718089aea9b9addfecace06cc21d2ddd  -- because I cannot save it as a patch set.

Sorry for the trouble -- but I was more or less finished with updating all the pages that have "register-true" -- because it was decided to change the name in the UI (and some of the labels in the dialog boxes).  Thanks.

Patch-set: 4
Reviewer: Gerrit User 1001613 <1001613@6622c302-437a-49ad-a91d-55ba7dcf1615>
Work-in-progress: false