Update patch set 2
Patch Set 2:

(1 comment)

Sorry to get you entangled in this -- but I need to check one more condition -- just to be sure.

(Based on experiment) I think this command (and help page) "Edit - Links to External Files" only applies to master documents (e.g., simply inserting hyperlinks into a document does not activate
the command in the menu). 

This is a different procedure than the one described in your test.

[ Was planning to ask others if this applies only to Master documents, but if you happen to know -- then I can finish this off. ]

Meanwhile, (Empirically from using UI) current functionality suggests that this warning is completely irrelevant (because the dialog for master documents is only to change which file is linked)?

(I have also discovered now that this warning was first added in the help file in 2010!)

In short, seems safe (even wise) to delete.

Patch-set: 2
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>