Update patch set 4
Patch Set 4: Code-Review-1

This version has the complete table for all the context menus in Navigator - and all the icons (except for Headings -- which I still can not find - from studying/searching in links.txt and from trying to find another help page that might have it -- have just used one that is close for now).

The only thing that remains to be done (besides the Headings icon) is to finish adding (and repairing) the links for the items in Headings context menus.

I think this approach is a good solution to the context menus for Navigator -- and now you can see what it would look like.

Assuming that comments, suggestions, objections will come if needed...

Patch-set: 4
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1