Update patch set 2
Patch Set 2:

(2 comments)

@Mike - a small, but critical, question about how the Document (field) is described. Please start with comment #1.

(all the other stuff in the patch can be ignored for this purpose)

And a little "background" to explain what motivates the question.

My recent bug reports about Document Fields and editing fields arise from trying to improve the documentation for Document fields. 

From my naive user POV, it was confusing that I could enter a ”Sender” Type (and select ”Company”) – assuming that I had just entered a ”Company” field in the document (i.e., only could display "company" from User Data) –  only to later discover that I could edit that field and change it to ”Email”, for example).   

(I think I have come to understand technically what is happening – in effect, ”Sender” type  is a variable field, where "Select" allows you to specify the ”variable name” of the field, and the program supplies the value.)  

But my query is not about the technical, but about finding a clear way to describe the situation for users who are reading the manual (which implies they are willing to try to learn), so trying to give a realistic chance for a thoughtful reader to form a useful understanding.  For example, not appreciating that the "field" Sender -> Company is only that relation until the same field is changed to "Sender --> Email" -- will presumably lead to unexpected surprises.

Patch-set: 2
Reviewer: Gerrit User 1000994 <1000994@6622c302-437a-49ad-a91d-55ba7dcf1615>
Work-in-progress: false