Update patch set 2
Patch Set 2:

(sorry for long reply -- but there is no "debate" or "disagreement" in what follows -- just seeking to move forward by asking particular questions of interpretation in relation to the current help text formulation, and conclude with a loose (but I hope more accurate/useful) description in next to last paragraph).

> my opinion is … create a better explanation of what the Type and Select 
> actually do … without trying to provide a rationale for an arbitrary design
> decision.

Amen.  That was (and still is) my objective, so we are presumably pulling the rope in the same direction.

> So I don't know which problem you try to solve here. 

Ok.  In this case, in relation to the help page, the general problem is (a) to accept the existing ‟insert field” implementation (dialog box and underlying software design) (i.e., not trying to propose new designs or changes here), and (b) to explain to a user (who may be familiar with creating and using variable fields in general, but not necessarily with Document fields in particular), about how to use the Insert Document dialog in an informed manner.

In relation to updating the help page, the ‟immediate” problem is to make sure that either (a) the existing description is accurate and comprehensible or (b) to make relevant improvement.

> Do I understand it correctly that the problem for you arises from the ability 
> to change the field later, to show a different data? I don't understand why is 
> that a problem actually...

Not to be philosophical, but to be precise, what exactly does ‟field” refer to here?  Is each individual ‟insert” a field, where that individual field has a ‟Type” property and a ‟Select” property?  Or should ‟field” be understood to refer to the ‟type” of field?

To illustrate concretely (in relation to the current help page): 

> Current text reads "Type Lists the available field types". You write "Type 
> lists the available document fields".

That is probably because I was (mis?)interpreting ‟field” to refer to the ‟type” of document information.  So if I insert a ‟Sender” field, then I can ‟Select” which (of the different kinds of) particular ‟Sender” information (from User Data) should be shown. 

And then reading the next sentence in the current description. 

> To add a field to your document, click a field type, click a field in the 
> <emph>Select </emph>list 

If I apply that current help page terminology (and using Type ‟Author” to illustrate), it says that I can select field Type ‟Author” and then click on field ‟Name” to insert the field ‟Name”.  (NB. I am just using the terminology of the help page here.)

And continuing, then this means that ‟Initials” is also a field.  
And finally (in that terminology) it must mean that I can insert an ‟Name” field and a ‟Initials” field (but then I can go into Edit Dialog) and change Name to Initials and vice versa.

Does that sound right to you?  (NB. my question here is from the POV of someone who is coming to the Insert Field – Document tab for first time, and is reading the help page to understand what is going on – and not from POV of someone who already understands the (somewhat arbitrary) meaning of Type and Select and their underlying logic/relation.)

As you helpfully pointed out: 

> The fact that later you can change the "Select" in the already inserted field, 
> but can't change the "Type", is just a random implementation decision, and is 
> not something fundamentally necessary: 

Exactly.  As long as you (the user) understand that ‟Sender” or ‟Author” describes a ‟category” where ‟any item” in the ‟Select” list can be assigned to that inserted field (where the system decides the content in all cases, drawing either on system data or values in User Data) -- then you will be a competent (and happy?) user of Insert - Fields - Document tab.

(probably my own terminology in the previous paragraph has to be improved, but if it approaches a more accurate conceptual model of ‟what is happening”, then I consider it a worthwhile improvement – and – if I am still misunderstanding – then this is also a sign that the help page needs to be improved. (-: )

Patch-set: 2
Reviewer: Gerrit User 1002268 <1002268@6622c302-437a-49ad-a91d-55ba7dcf1615>